+++ streamconnector/featureserializerv1.cpp
@@ -57,7 +57,7 @@ bool FeatureSerializerV1::store(IlwisObject *obj, const IOOptions &options)
         domainStreamer->store( fcoverage->attributeDefinitions().domain().ptr(), options);
 
         std::vector<QString> indexes = fcoverage->attributeDefinitions().indexes();
-        _stream << indexes.size();
+        _stream << (quint32)indexes.size();
         for(auto index : indexes)
             _stream << index;
     }else{
@@ -129,7 +129,7 @@ bool FeatureSerializerV1::loadMetaData(IlwisObject *obj, const IOOptions &option
             return false;
         IDomain dom(type | valueType);
         domainStreamer->loadMetaData( dom.ptr(), options);
-        size_t nrOfVariants;
+        quint32 nrOfVariants;
         _stream >> nrOfVariants;
         std::vector<QString> variants(nrOfVariants);
         for(int i =0; i < nrOfVariants; ++i){
